refactor(api): move csv helper to src/shared
