+++ api/lib/infrastructure/serializers/csv/csv-serializer.js
@@ -6,9 +6,9 @@ import {
   headers,
 } from '../../../../src/certification/shared/infrastructure/utils/csv/sessions-import.js';
 import { FileValidationError } from '../../../../src/shared/domain/errors.js';
+import { csvHelper } from '../../../../src/shared/infrastructure/helpers/csv.js';
 import { convertDateValue } from '../../../../src/shared/infrastructure/utils/date-utils.js';
 import { logger } from '../../../../src/shared/infrastructure/utils/logger.js';
-import { csvHelper } from '../../helpers/csv.js';
 
 const { isEmpty } = lodash;
