--- api/tests/unit/infrastructure/helpers/files/organizations-empty-file.csv
@@ -1,1 +1,0 @@
-
