--- api/tests/unit/infrastructure/helpers/files/withHeader-test.csv
@@ -1,5 +1,0 @@
-uai,name
-0080017a ,Collège Les Pixous
-
-0080018b ,Lycée Pix
-0080040a   ,Lycée Tant Pix
